syntax: extract `builin_attrs.rs`.
