--- src/libsyntax/feature_gate/attr.rs
@@ -1,1 +1,0 @@
-//! TODO
